Refactor toolbar using custom enum
